Align optimisation result
